backend/requirements.txt
# Core
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# Celery
celery==5.4.0
django-celery-beat==2.6.0
redis==5.0.4

# Storage (S3/MinIO)
boto3==1.34.84
django-storages==1.14.3

# Payments
razorpay==1.4.2

# Utilities
Pillow==10.3.0
python-decouple==3.8
gunicorn==22.0.0
whitenoise==6.6.0
django-extensions==3.2.3
phonenumbers==8.13.34

# Monitoring
sentry-sdk==1.45.0

# API docs
drf-spectacular==0.27.2

# KYC / HTTP
httpx==0.27.0
